+++ packages/atomic-angular/projects/atomic-angular/src/lib/stencil-generated/atomic-angular.module.ts
@@ -1,0 +1,152 @@
+/* tslint:disable */
+/* auto-generated angular module */
+import {CommonModule} from '@angular/common';
+import {ModuleWithProviders, NgModule} from '@angular/core';
+
+        
+import {
+AtomicBreadbox,
+AtomicCategoryFacet,
+AtomicColorFacet,
+AtomicComponentError,
+AtomicDidYouMean,
+AtomicExternal,
+AtomicFacet,
+AtomicFacetManager,
+AtomicFieldCondition,
+AtomicFormatCurrency,
+AtomicFormatNumber,
+AtomicFormatUnit,
+AtomicFrequentlyBoughtTogether,
+AtomicIcon,
+AtomicLoadMoreResults,
+AtomicNoResults,
+AtomicNumericFacet,
+AtomicNumericRange,
+AtomicPager,
+AtomicQueryError,
+AtomicQuerySummary,
+AtomicRatingFacet,
+AtomicRatingRangeFacet,
+AtomicRefineModal,
+AtomicRefineToggle,
+AtomicRelevanceInspector,
+AtomicResult,
+AtomicResultBadge,
+AtomicResultDate,
+AtomicResultFieldsList,
+AtomicResultIcon,
+AtomicResultImage,
+AtomicResultLink,
+AtomicResultList,
+AtomicResultMultiValueText,
+AtomicResultNumber,
+AtomicResultPrintableUri,
+AtomicResultRating,
+AtomicResultSectionActions,
+AtomicResultSectionBadges,
+AtomicResultSectionBottomMetadata,
+AtomicResultSectionEmphasized,
+AtomicResultSectionExcerpt,
+AtomicResultSectionTitle,
+AtomicResultSectionTitleMetadata,
+AtomicResultSectionVisual,
+AtomicResultTemplate,
+AtomicResultText,
+AtomicResultsPerPage,
+AtomicSearchBox,
+AtomicSearchBoxQuerySuggestions,
+AtomicSearchBoxRecentQueries,
+AtomicSearchInterface,
+AtomicSortDropdown,
+AtomicSortExpression,
+AtomicTableElement,
+AtomicText,
+AtomicTimeframe,
+AtomicTimeframeFacet
+} from './components';
+
+        
+import {defineCustomElements} from '@coveo/atomic/loader';
+defineCustomElements(window);
+
+        
+const DECLARATIONS = [
+AtomicBreadbox,
+AtomicCategoryFacet,
+AtomicColorFacet,
+AtomicComponentError,
+AtomicDidYouMean,
+AtomicExternal,
+AtomicFacet,
+AtomicFacetManager,
+AtomicFieldCondition,
+AtomicFormatCurrency,
+AtomicFormatNumber,
+AtomicFormatUnit,
+AtomicFrequentlyBoughtTogether,
+AtomicIcon,
+AtomicLoadMoreResults,
+AtomicNoResults,
+AtomicNumericFacet,
+AtomicNumericRange,
+AtomicPager,
+AtomicQueryError,
+AtomicQuerySummary,
+AtomicRatingFacet,
+AtomicRatingRangeFacet,
+AtomicRefineModal,
+AtomicRefineToggle,
+AtomicRelevanceInspector,
+AtomicResult,
+AtomicResultBadge,
+AtomicResultDate,
+AtomicResultFieldsList,
+AtomicResultIcon,
+AtomicResultImage,
+AtomicResultLink,
+AtomicResultList,
+AtomicResultMultiValueText,
+AtomicResultNumber,
+AtomicResultPrintableUri,
+AtomicResultRating,
+AtomicResultSectionActions,
+AtomicResultSectionBadges,
+AtomicResultSectionBottomMetadata,
+AtomicResultSectionEmphasized,
+AtomicResultSectionExcerpt,
+AtomicResultSectionTitle,
+AtomicResultSectionTitleMetadata,
+AtomicResultSectionVisual,
+AtomicResultTemplate,
+AtomicResultText,
+AtomicResultsPerPage,
+AtomicSearchBox,
+AtomicSearchBoxQuerySuggestions,
+AtomicSearchBoxRecentQueries,
+AtomicSearchInterface,
+AtomicSortDropdown,
+AtomicSortExpression,
+AtomicTableElement,
+AtomicText,
+AtomicTimeframe,
+AtomicTimeframeFacet
+]
+
+        
+@NgModule({
+  declarations: DECLARATIONS,
+  exports: DECLARATIONS,
+  providers: [],
+  imports: [CommonModule],
+})
+export class AtomicAngularModule {
+  static forRoot(): ModuleWithProviders<AtomicAngularModule> {
+    return {
+      ngModule: AtomicAngularModule,
+      providers: [],
+    };
+  }
+}
+
+