+++ packages/atomic/cypress/integration/query-summary-selectors.ts
@@ -7,5 +7,6 @@ export const QuerySummarySelectors = {
   placeholder: () =>
     QuerySummarySelectors.shadow().find('[part="placeholder"]'),
   container: () => QuerySummarySelectors.shadow().find('[part="container"]'),
+  duration: () => QuerySummarySelectors.shadow().find('[part="duration"]'),
   liveRegion: () => AriaLiveSelectors.region('query-summary'),
 };