+++ packages/quantic/cypress/integration/result-template/result-template-actions.ts
@@ -7,7 +7,7 @@ const resultTemplateActions = (selector: ResultTemplateSelector) => {
   return {
     appendChildren: (
       tagElement: string,
-      props: any = {},
+      props: Record<string, string> = {},
       innerHtml?: string
     ) => {
       selector